requirements.txt
RPi.GPIO==0.7.1
numpy==1.24.3
pytest==7.4.0 